feat(dotnet): add support for proxy config (#2092)
* feat(dotnet): add support for proxy config

* lint

* update

* lint

* update

* lint

* update

+++ statsig-ffi/include/statsig_ffi.h
@@ -309,7 +309,7 @@ uint64_t statsig_options_create(const char *specs_url,
                                 SafeOptBool fallback_to_statsig_api,
                                 SafeOptBool use_third_party_ua_parser,
                                 const char *proxy_host,
-                                uint16_t proxy_port,
+                                int proxy_port,
                                 const char *proxy_auth,
                                 const char *proxy_protocol);
